requirements.txt
requests>=2.31.0
plotly>=5.17.0
folium>=0.14.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
plotnine>=0.12.0

